+++ vhost-device-vsock/CHANGELOG.md
@@ -7,6 +7,8 @@
 
 ### Fixed
 
+- [#936](https://github.com/rust-vmm/vhost-device/pull/936) vsock: Access virtqueues only if they are ready
+
 ### Deprecated
 
 ## v0.3.0